Add Android Archive (.aar) support to build server
Many Android-specific libraries are moving toward the use of Android
Archive (.aar) files rather than Java Archive (.jar) files due to
being able to include additional Android resources, specify the minSdk
required for the library, etc. This commit adds support for subsets of
the Android Archive functionality to the App Inventor build server to
allow for more advanced components to include .aar files as
dependencies.

One caveat is that the archive's internal classes.jar file must be
extracted so that it can be included in the build.xml classpaths
because Ant does not know how to consume .aar for classpath elements.

Change-Id: I19bc62a69d82c7d51585d8d2acd2bb39ad6f9716