Fix empty dialog box for logged in user after ant clean
During development if one is logged in to App Inventor and performs a
clean rebuild, the account no longer exists. This results in a dialog
being presented that is empty and HTTP 401 Unauthorized. This commit
detects the empty dialog and redirects the user to the login page to
have the App Inventor cookie reset.

Change-Id: I0327cdb34bb372b9055c2b17d8c9d054549067ab